add utils for Joi & update methods
