+++ controllers/contacts/index.js
@@ -1,4 +1,4 @@
-const listContacts = require('./listContacts');
+const getContacts = require('./listContacts');
 const getContactById = require('./getContactById');
 const addContact = require('./addContact');
 const removeContact = require('./removeContact');
@@ -6,7 +6,7 @@ const updateContact = require('./updateContact');
 const updateFavorite = require('./updateFavorite');
 
 module.exports = {
-    listContacts,
+    getContacts,
     getContactById,
     addContact,
     removeContact,
