+++ models/contact.js
@@ -1,5 +1,4 @@
 const {Schema, model} = require('mongoose');
-const Joi = require('joi');
 
 const contactSchema = Schema({
     name: {
@@ -17,17 +16,6 @@ const contactSchema = Schema({
         default: false,
       },
 });
-
-const joiSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().required(),
-  phone: Joi.string().required(),
-  favorite: Joi.bool()
-})
-
-const favoriteJoiSchema = Joi.object({
-  favorite: Joi.bool().required()
-})
 
 const Contact = model('contact', contactSchema);
